fix: 🐛 set case status
approved when decision and payment exists, partially approved when
decision has some declined and payment exists, rejected when decision is
'Avslag', processing when only calculation exists
